Hide select element
This is a bit challenging to reproduce because the offset is set to `-999em`, but I am seeing the unstyled `select` element in my app when I have several tabs that are very wide.

Any reason we should not just hide the element? I seem to still be getting events ok when I manually set this style.